+++ services/qa/index.js
@@ -1,0 +1,3 @@
+const { runQaFixtures } = require('./fixture-runner');
+
+module.exports = { runQaFixtures };